+++ app/src/routes/_mainLayout/_static/about.tsx
@@ -1,4 +1,3 @@
-import UnderConstruction from "@/components/under-construction";
 import { createFileRoute } from "@tanstack/react-router";
 
 export const Route = createFileRoute("/_mainLayout/_static/about")({
@@ -13,9 +12,4 @@ export const Route = createFileRoute("/_mainLayout/_static/about")({
             },
         ],
     }),
-    component: RouteComponent,
 });
-
-function RouteComponent() {
-    return <UnderConstruction />;
-}
