+++ app/src/routes/_mainLayout/_static/contact-us.lazy.tsx
@@ -1,0 +1,10 @@
+import ContactSection from '@/components/store/landing/contact-section';
+import { createLazyFileRoute } from '@tanstack/react-router'
+
+export const Route = createLazyFileRoute('/_mainLayout/_static/contact-us')({
+    component: RouteComponent,
+})
+
+function RouteComponent() {
+    return <ContactSection />;
+}
