+++ app/src/routes/_mainLayout/_static/contact-us.tsx
@@ -1,4 +1,3 @@
-import ContactSection from '@/components/store/landing/contact-section';
 import { createFileRoute } from '@tanstack/react-router'
 
 export const Route = createFileRoute('/_mainLayout/_static/contact-us')({
@@ -13,9 +12,4 @@ export const Route = createFileRoute('/_mainLayout/_static/contact-us')({
             },
         ],
     }),
-    component: RouteComponent,
 })
-
-function RouteComponent() {
-    return <ContactSection />;
-}
